refactor createQuestionOrderedInfoList to static method in CohortRecord
